Validate in the renderer getConvertedValue method.
