+++ src/main/java/org/primefaces/extensions/component/inputphone/InputPhoneRenderer.java
@@ -15,22 +15,28 @@
  */
 package org.primefaces.extensions.component.inputphone;
 
+import com.google.i18n.phonenumbers.NumberParseException;
+import com.google.i18n.phonenumbers.PhoneNumberUtil;
+import com.google.i18n.phonenumbers.Phonenumber;
 import java.io.IOException;
 import java.util.Arrays;
 import java.util.Collection;
-
 import javax.faces.FacesException;
+import javax.faces.application.FacesMessage;
 import javax.faces.component.UIComponent;
 import javax.faces.context.FacesContext;
 import javax.faces.context.ResponseWriter;
-
+import javax.faces.convert.Converter;
+import javax.faces.convert.ConverterException;
 import org.apache.commons.lang3.StringUtils;
 import org.primefaces.component.inputtext.InputText;
+import org.primefaces.extensions.util.MessageFactory;
 import org.primefaces.json.JSONArray;
 import org.primefaces.renderkit.InputRenderer;
 import org.primefaces.util.ComponentUtils;
 import org.primefaces.util.Constants;
 import org.primefaces.util.HTML;
+import org.primefaces.util.LangUtils;
 import org.primefaces.util.WidgetBuilder;
 
 /**
@@ -40,6 +46,8 @@ import org.primefaces.util.WidgetBuilder;
  * @since 7.0
  */
 public class InputPhoneRenderer extends InputRenderer {
+
+    private static final String MESSAGE_INVALID_KEY = "primefaces.extensions.inputphone.INVALID";
 
     @Override
     public void decode(FacesContext context, UIComponent component) {
@@ -71,6 +79,41 @@ public class InputPhoneRenderer extends InputRenderer {
 
         encodeMarkup(context, inputPhone, valueToRender);
         encodeScript(context, inputPhone);
+    }
+
+    @Override
+    public Object getConvertedValue(FacesContext context, UIComponent component, Object submittedValue) throws ConverterException {
+        final String value = (String) submittedValue;
+        if (LangUtils.isValueBlank(value)) {
+            return null;
+        }
+
+        final InputPhone inputPhone = (InputPhone) component;
+        final Converter converter = inputPhone.getConverter();
+
+        if (converter != null) {
+            return converter.getAsObject(context, inputPhone, value);
+        }
+
+        String country = context.getExternalContext().getRequestParameterMap().get(inputPhone.getClientId() + "_iso2");
+        if (country == null || InputPhone.COUNTRY_AUTO.equals(country)) {
+            country = Constants.EMPTY_STRING;
+        }
+        try {
+            final PhoneNumberUtil phoneNumberUtil = PhoneNumberUtil.getInstance();
+            final Phonenumber.PhoneNumber phoneNumber = phoneNumberUtil.parse(value, country);
+            if (!phoneNumberUtil.isValidNumber(phoneNumber)) {
+                throw new ConverterException(getMessage());
+            }
+            return value;
+        }
+        catch (final NumberParseException e) {
+            throw new ConverterException(getMessage());
+        }
+    }
+
+    protected FacesMessage getMessage() {
+        return MessageFactory.getMessage(MESSAGE_INVALID_KEY, FacesMessage.SEVERITY_ERROR);
     }
 
     protected void encodeMarkup(FacesContext context, InputPhone inputPhone, String valueToRender) throws IOException {
